requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
requests>=2.28.1
beautifulsoup4>=4.11.1
nltk>=3.7
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.28.1
pydantic>=1.10.2
python-dotenv>=0.21.0
Jinja2>=3.0.0
python-multipart
langchain>=0.0.200
